# Conflicts:
#	pom.xml

+++ src/test/java/net/openhft/chronicle/bytes/MappedFileTest.java
@@ -40,6 +40,12 @@ public class MappedFileTest {
     public void checkThreadDump() {
         threadDump.assertNoNewThreads();
     }
+
+    @Test
+    public void testWarmup() throws InterruptedException {
+        MappedFile.warmup();
+    }
+
     @Ignore("todo fix sometimes fails on TC")
     @Test
     public void testReferenceCounts() throws IOException {
@@ -101,6 +107,9 @@ public class MappedFileTest {
 
     @Test
     public void largeReadOnlyFile() throws IOException {
+        if (Runtime.getRuntime().totalMemory() < Integer.MAX_VALUE)
+            return;
+
         File file = File.createTempFile("largeReadOnlyFile", "deleteme");
         file.deleteOnExit();
         try (MappedBytes bytes = MappedBytes.mappedBytes(file, 1 << 30, OS.pageSize())) {
